\name{dbeOutput-class}
\docType{class}
\alias{dbeOutput}
\alias{dbeOutput-class}
\title{Class "dbeOutput"}
\description{Outcome object from \emph{COSTdbe} methods}

\section{Slots}{
\describe{
    \item{\code{desc}:}{Object of class \code{"character"}. Descriptive slot}
    \item{\code{species}:}{Object of class \code{"character"}. Species}
    \item{\code{catchCat}:}{Object of class \code{"character"}. Catch category}
    \item{\code{param}:}{Object of class \code{"character"}. Estimated parameter (eg "maturity","sex-ratio",...)}
    \item{\code{strataDesc}:}{Object of class \code{"strIni"}. Considered stratification}
    \item{\code{methodDesc}:}{Object of class \code{"character"}. Used method (eg "analytical","bootstrap" or "Bayesian")}
    \item{\code{nSamp}:}{Object of class \code{"list"}. Number of samples\\
\tabular{llll}{
 \tab len \tab \code{data.frame} \tab Number of length samples. \cr
 \tab age \tab \code{data.frame} \tab Number of age samples. \cr
}}
    \item{\code{nMeas}:}{Object of class \code{"list"}. Number of individual measured\\
\tabular{llll}{
 \tab len \tab \code{data.frame} \tab Number of individual measured for length data. \cr
 \tab age \tab \code{data.frame} \tab Number of individual measured for age data. \cr
}}
    \item{\code{lenStruc}:}{Object of class \code{"list"}. Estimates of the length structure\\
\tabular{llll}{
 \tab estim \tab \code{data.frame} \tab Final estimates (\emph{length} field added). \cr
 \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{length} and \emph{iter} fields added). \cr
}}
    \item{\code{lenVar}:}{Object of class \code{"data.frame"}. Variance estimates of 'lenStruc'}
    \item{\code{lenNum}:}{Object of class \code{"list"}. Precision estimates for length structure\\
\tabular{llll}{
 \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
 \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
 \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
}}    
    \item{\code{ageStruc}:}{Object of class \code{"list"}. Estimates of the age structure\\
\tabular{llll}{
 \tab estim \tab \code{data.frame} \tab Final estimates (\emph{age} field added). \cr
 \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{age} and \emph{iter} fields added). \cr
}}    
    \item{\code{ageVar}:}{Object of class \code{"data.frame"}. Variance estimates of 'ageStruc'}
    \item{\code{ageNum}:}{Object of class \code{"list"}. Precision estimates for age structure\\
\tabular{llll}{
 \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
 \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
 \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
}}    
    \item{\code{totalN}:}{Object of class \code{"list"}. Total number estimates\\
\tabular{llll}{
 \tab estim \tab \code{data.frame} \tab Final estimates. \cr
 \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{iter} field added). \cr
}}    
    \item{\code{totalNvar}:}{Object of class \code{"data.frame"}. Variance estimates of 'totalN'}
    \item{\code{totalNnum}:}{Object of class \code{"list"}. Precision estimates for total numbers\\
\tabular{llll}{
 \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
 \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
 \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
}}    
    \item{\code{totalW}:}{Object of class \code{"list"}. Total weight estimates\\
\tabular{llll}{
 \tab estim \tab \code{data.frame} \tab Final estimates. \cr
 \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{iter} field added). \cr
}}    
    \item{\code{totalWvar}:}{Object of class \code{"data.frame"}. Variance estimates of 'totalW'}
    \item{\code{totalWnum}:}{Object of class \code{"list"}. Precision estimates for total weights\\
\tabular{llll}{
 \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
 \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
 \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
}}    
}
}

%\tabular{lllll}{
%\bold{slot} \tab \bold{desc} \tab \bold{elements} \tab \bold{class} \tab \bold{description} \cr
%\bold{\code{desc}} \tab \tab \tab \code{character} \tab Descriptive slot \cr
%\bold{\code{species}} \tab \tab \tab \code{character} \tab Species \cr
%\bold{\code{catchCat}} \tab \tab \tab \code{character} \tab Catch category (eg "LAN", "DIS" or "all") \cr
%\bold{\code{param}} \tab \tab \tab \code{character} \tab Parameter estimated (eg "maturity","sex-ratio",...)  \cr
%\bold{\code{strataDesc}} \tab \tab \tab \code{strIni} \tab Stratification considered \cr
%\bold{\code{methodDesc}} \tab \tab \tab \code{character} \tab Used method (eg "analytical","bootstrap" or "Bayesian") \cr
%}
%All the following slots contain numerics or dataframes that will be generated by COSTdbe methods :\cr
%\tabular{lllll}{
%\bold{nSamp} \tab \tab \tab \code{list} \tab Number of samples\cr
% \tab \tab len \tab \code{data.frame} \tab Number of length samples. \cr
% \tab \tab age \tab \code{data.frame} \tab Number of age samples. \cr
%\bold{nMes} \tab \tab \tab \code{list} \tab Number of individual measured\cr
% \tab \tab len \tab \code{data.frame} \tab Number of individual measured for length data. \cr
% \tab \tab age \tab \code{data.frame} \tab Number of individual measured for age data. \cr
%\bold{lenStruc} \tab \tab \tab \code{list} \tab Estimates of the length structure\cr
% \tab \tab estim \tab \code{data.frame} \tab Final estimates (\emph{length} field added). \cr
% \tab \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{length} and \emph{iter} fields added). \cr
%\bold{lenVar} \tab \tab \tab \code{data.frame} \tab Estimates of the variance of 'lenStruc'\cr
%\bold{lenNum} \tab \tab \tab \code{list} \tab Precision estimates for length structure\cr
% \tab \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
% \tab \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
% \tab \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
%\bold{ageStruc} \tab \tab \tab \code{list} \tab Estimates of the age structure\cr
% \tab \tab estim \tab \code{data.frame} \tab Final estimates (\emph{age} field added). \cr
% \tab \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{age} and \emph{iter} fields added). \cr
%\bold{ageVar} \tab \tab \tab \code{data.frame} \tab Estimates of the variance of 'ageStruc'\cr
%\bold{ageNum} \tab \tab \tab \code{list} \tab Precision estimates for age structure\cr
% \tab \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
% \tab \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
% \tab \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
%\bold{totalN} \tab \tab \tab \code{list} \tab Estimates of the total number of the parameters\cr
% \tab \tab estim \tab \code{data.frame} \tab Final estimates. \cr
% \tab \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{iter} field added). \cr
%\bold{totalNvar} \tab \tab \tab \code{data.frame} \tab Estimates of the variance of 'totalN'\cr
%\bold{totalNnum} \tab \tab \tab \code{list} \tab Precision estimates for total numbers\cr
% \tab \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
% \tab \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
% \tab \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
%\bold{totalW} \tab \tab \tab \code{list} \tab Estimates of the total weight of the parameters\cr
% \tab \tab estim \tab \code{data.frame} \tab Final estimates. \cr
% \tab \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{iter} field added). \cr
%\bold{totalWvar} \tab \tab \tab \code{data.frame} \tab Estimates of the variance of 'totalW'\cr
%\bold{totalWnum} \tab \tab \tab \code{list} \tab Precision estimates for total weights\cr
% \tab \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
% \tab \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
% \tab \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
%}}

\author{Mathieu Merzereaud}

\seealso{\code{\link{totVolume}}, \code{\link{dbePlot}}
}

\examples{
showClass("dbeOutput")
}

\keyword{classes}

