\name{LEM.objects}
\alias{LEM.CE}
\alias{LEM.CEcon}
\alias{LEM.CL}
\alias{LEM.CLcon}
\alias{LEM.CS}
\alias{LEM.CScon}
\alias{LEM.strat}

\docType{data}

\title{ToDo}

\description{
ToDo}

\usage{LEM.CE}

\format{ToDo}

\section{WARNING}{
This data is provided is for testing only and should not be used for any other purpose. 
}

\details{
ToDo
}

\source{
ToDo
}

\examples{
data(LEMexample)
}

\keyword{datasets}

