\name{plus-dbeOutput}
\docType{methods}
\alias{+,dbeOutput,dbeOutput-method}
\title{Addition of `dbeOutput' objects}
\description{This method implements addition of 'dbeOutput' objects provided by COSTdbe package.}
\section{Methods}{
\describe{
\item{}{\code{signature(e1 = "dbeOutput", e2 = "dbeOutput")}}
}}
\details{
Every dataframe are summed, except \emph{"num"} elements that are set to NA. 
}
\keyword{methods}

