\name{RaiseLgth}
\alias{RaiseLgth}
\alias{RaiseLgth,dbeOutput,csDataCons,clDataCons-method}
\docType{methods}
\title{Estimation of total numbers-at-length from market sampling}
\description{
ToDo
}

\usage{
RaiseLgth(dbeOutput,csObject,clObject,spp,taxon,sex=as.character(NA),\dots)
}

\arguments{
  \item{dbeOutput}{A \emph{dbeOutput} object.}
  \item{csObject}{A \emph{csDataCons} object matching 'dbeOutput' specifications.}
  \item{clObject}{A \emph{clDataCons} object matching 'dbeOutput' specifications.}
  \item{spp}{Species}
  \item{taxon}{Taxon}
  \item{sex}{Sex}
  \item{\dots}{Further arguments}  
}

\references{ToDo}

\value{An updated object of class \emph{dbeOutput}.}


\author{Mathieu Merzereaud}
\seealso{\code{\link{dbeOutput}}, \code{\link{dbeObject}}, \code{\link[COSTcore]{csDataCons}}, \code{\link[COSTcore]{clDataCons}}
}

\examples{

}
\keyword{methods}
